+++ src/Resources/contao/languages/en/modules.php
@@ -27,3 +27,4 @@ $GLOBALS['TL_LANG']['FMD']['wertungsportal_turnier'] = array('Tournament search'
 $GLOBALS['TL_LANG']['FMD']['wertungsportal_bestenliste'] = array('DWZ best list', 'DWZ best list (top x, all players or women only) from the local persons table');
 $GLOBALS['TL_LANG']['FMD']['wertungsportal_token'] = array('Interface registration', 'Form for clubs to request an access key for the club list interface');
 $GLOBALS['TL_LANG']['MOD']['wp-referenten'] = array('Rating officers', 'Manage the rating officers of the associations');
+$GLOBALS['TL_LANG']['FMD']['wertungsportal_referenten'] = array('Rating officers','The rating officers of the associations as a hierarchy, with address and protected e-mail');